+++ build_emulationstation.sh
@@ -21,24 +21,30 @@ fi
 NAME=`echo ${NAME} | tr '[:lower:]' '[:upper:]'`
 echo "export softname=\"dArkOS-${NAME}\"" | sudo tee -a Arkbuild/home/ark/ES_VARIABLES.txt
 
-call_chroot "apt-get -y update && eatmydata apt-get -y install libfreeimage3 fonts-droid-fallback libfreetype6 curl vlc-bin libsdl2-mixer-2.0-0"
-call_chroot "cd /home/ark &&
-  source ES_VARIABLES.txt &&
-  rm ES_VARIABLES.txt &&
-  git clone --recursive --depth=1 https://github.com/christianhaitian/EmulationStation-fcamod -b ${ES_BRANCH} &&
-  cd EmulationStation-fcamod &&
-  git submodule update --init &&
-  for f in \$(find . -type f \( -name '*.cpp' -o -name '*.h' \) -exec grep -L '<string>' {} \;); do
-    sed -i '1i#include <string>' \"\$f\";
-  done &&
-  sed -i '1i#include <ctime>' es-core/src/utils/TimeUtil.h &&
-  cmake -DSCREENSCRAPER_DEV_LOGIN=\"devid=\$devid&devpassword=\$devpass\" -DGAMESDB_APIKEY=\"\$apikey\" -DSCREENSCRAPER_SOFTNAME=\"\$softname\" . &&
-  make -j\$(nproc) &&
-  mkdir -pv /usr/bin/emulationstation &&
-  cp -a emulationstation /usr/bin/emulationstation &&
-  chmod 777 /usr/bin/emulationstation &&
-  cp -a resources /usr/bin/emulationstation/
-  "
+if [ -f "Arkbuild_package_cache/${CHIPSET}/emulationstation_${ES-BRANCH}.tar.gz" ] && [ "$(cat Arkbuild_package_cache/${CHIPSET}/emulationstation_${ES-BRANCH}.commit)" == "$(curl -s https:/api.github.com/repos/christianhaitian/EmulationStation-fcamod/commits/${ES-BRANCH} | jq -r '.sha')" ]; then
+    sudo tar -xvzpf Arkbuild_package_cache/${CHIPSET}/emulationstation_${ES-BRANCH}.tar.gz
+else
+	call_chroot "apt-get -y update && eatmydata apt-get -y install libfreeimage3 fonts-droid-fallback libfreetype6 curl vlc-bin libsdl2-mixer-2.0-0"
+	call_chroot "cd /home/ark &&
+	  source ES_VARIABLES.txt &&
+	  rm ES_VARIABLES.txt &&
+	  git clone --recursive --depth=1 https://github.com/christianhaitian/EmulationStation-fcamod -b ${ES_BRANCH} &&
+	  cd EmulationStation-fcamod &&
+	  git submodule update --init &&
+	  for f in \$(find . -type f \( -name '*.cpp' -o -name '*.h' \) -exec grep -L '<string>' {} \;); do
+		sed -i '1i#include <string>' \"\$f\";
+	  done &&
+	  sed -i '1i#include <ctime>' es-core/src/utils/TimeUtil.h &&
+	  cmake -DSCREENSCRAPER_DEV_LOGIN=\"devid=\$devid&devpassword=\$devpass\" -DGAMESDB_APIKEY=\"\$apikey\" -DSCREENSCRAPER_SOFTNAME=\"\$softname\" . &&
+	  make -j\$(nproc) &&
+	  mkdir -pv /usr/bin/emulationstation &&
+	  cp -a emulationstation /usr/bin/emulationstation &&
+	  chmod 777 /usr/bin/emulationstation &&
+	  cp -a resources /usr/bin/emulationstation/
+	  "
+	  sudo tar -czpf Arkbuild_package_cache/${CHIPSET}/emulationstation_${ES-BRANCH}.tar.gz Arkbuild/usr/bin/emulationstation/
+	  sudo git --git-dir=Arkbuild/home/ark/${CHIPSET}_core_builds/EmulationStation-fcamod/.git --work-tree=Arkbuild/home/ark/${CHIPSET}_core_builds/EmulationStation-fcamod rev-parse HEAD > Arkbuild_package_cache/${CHIPSET}/emulationstation_${ES-BRANCH}.commit
+fi
 sudo rm -rf Arkbuild/home/ark/EmulationStation-fcamod
 sudo mkdir -p Arkbuild/etc/emulationstation/themes
 if [[ "${BUILD_ARMHF}" == "y" ]]; then
